requirements.txt
google-generativeai
streamlit
python-docx